Animation improvements and unit tests added.

+++ BorisEngine2/Renderable.h
@@ -5,10 +5,10 @@
 
 struct Colour
 {
-	Uint8 R;
-	Uint8 G;
-	Uint8 B;
-	Uint8 A;
+	int R;
+	int G;
+	int B;
+	int A;
 	bool operator==(const Colour& other) const
 	{
 		return R == other.R && G == other.G && B == other.B && A == other.A;
